chore: Non-functional refactor
