+++ src/spectre_core/receivers/plugins/_signal_generator.py
@@ -20,7 +20,8 @@ from spectre_core.capture_configs import (
 from ._signal_generator_gr import cosine_wave, tagged_staircase
 from ._gr import capture
 from ._receiver_names import ReceiverName
-from .._receiver import SpecName, Specs, Receiver
+from .._receiver import Receiver
+from .._specs import Specs, SpecName
 from .._register import register_receiver
